+++ shardingsphere-scaling/shardingsphere-scaling-bootstrap/src/main/java/org/apache/shardingsphere/scaling/ServerBootstrap.java
@@ -29,7 +29,7 @@ import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shardingsphere.scaling.core.config.ScalingContext;
-import org.apache.shardingsphere.scaling.utils.ScalingConfigUtil;
+import org.apache.shardingsphere.scaling.util.ScalingConfigUtil;
 import org.apache.shardingsphere.scaling.web.HttpServerInitializer;
 
 /**
