+++ shardingsphere-scaling/shardingsphere-scaling-bootstrap/src/test/java/org/apache/shardingsphere/scaling/web/HttpServerHandlerTest.java
@@ -32,8 +32,8 @@ import lombok.SneakyThrows;
 import org.apache.shardingsphere.scaling.core.config.ScalingContext;
 import org.apache.shardingsphere.scaling.core.config.ServerConfiguration;
 import org.apache.shardingsphere.scaling.core.execute.engine.ShardingScalingExecuteEngine;
-import org.apache.shardingsphere.scaling.utils.ReflectionUtil;
-import org.apache.shardingsphere.scaling.utils.ScalingConfigurationUtil;
+import org.apache.shardingsphere.scaling.util.ReflectionUtil;
+import org.apache.shardingsphere.scaling.util.ScalingConfigurationUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
